Refactor fetch interview request and update local state on cancel interview
